Move project deletion to util
